agent/requirements.txt
anthropic==0.28.0
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
gitpython==3.1.43
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.32.3
eventlet==0.36.1
